requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
channels==4.1.0
constantly==23.10.4
cryptography==43.0.3
daphne==4.1.2
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
grpcio==1.66.0
grpcio-tools==1.66.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
pika==1.3.2
protobuf==5.27.3
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==24.2.1
python-decouple==3.8
service-identity==24.2.0
sqlparse==0.5.1
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
zope.interface==7.1.1
